Initial, Linux-only, plugin
The plan is to extend the Makefile to cover OS X and Windows, both under Visual Studio and under MinGW (picking whichever one it finds first).

Stata does not exit with an error code when things fail, which means testing is going to be a tedious case of recording expected logfiles.