Rather than escaping Stata's spaces, just quote.
Doublequoting is both bash and cmd compatible.
(to use this, I have to disable the space-escaping because cmd doesn't
understand backslash escapes; wildcard still needs it though so I left it
as a make macro)

This almost makes this compatible with both shells available for Windows,
unfortunately, mixing shells is having side-effects:
 - make test leaves you in a subshell??
 - RM, CP and CAT break (because del, copy and type are cmd builtins)
so it's really not worth chasing this rabbit.
if make had cp and rm built-in then it could use its "is it Wandows?"
logic to help here, but it doesn't, so I will stop now. I did want to
record the tip that doublequoting *in recipes* works (just not with
$(wildcard))