Instead of deps/, just use reference/ to hold external libraries, in addition to papers and such already there.
I want all dependencies to come in externally, via system package managers or at least standard manual instructions.
This build must *not* be dependent on my particular setup: we must be able to quickly set up a new build machine.
The longest step should be waiting for the compiler to install.