svm_export, which saves a
As this duplicates what the #if DEBUG did in train(), that is now gone.
You can always ad-hoc it back if you really need to compare the difference
in states between training and, say, predicting, but a trained model *should* be immutable.

There is a *single global model variable* now, for reasons explained in the comments.