Thread the commands together
It turns out that what passes for polymorphism in Stata is this:
 ereturn local predict = "<prediction command>"
i.e. at the end of a fit, store the (string!) name of the function
that should handle prediction into a well known global
i.e. something you'd do in bash
But whatever floats your banana boat I guess. choo choo choo.

Now that I know this, I don't need subcommands that badly,
so 'svm' is an alias for 'svm_train'.

svm_export and svm_import are still around, but svm_import doesn't
know to set predict, and it isn't clear to me that that's a good idea,
seeing as loading a model without the identical original data
(remember libsvm's column IDs don't map to the Stata ones unless you
 always precisely do "svm_train *")