First draft of packaging for live Stata consumption.
I don't have a cross-platform buildbot so there necessarily will be
some manual effort to get a release out, on top of the normal manual effort,
but I envision automating as much as possible. My build vision is:

for each target architecture:
  log into the host machine with that arch (or activate the appropriate cross-compiler)
    make dist
    upload build/ to build/ on the staging server (e.g. the OS X build machine, say)
on the staging server:
  make pkg  -> produces svm/{bin/{WIN,MACINTEL,LINUX,...},*ado,*sthlp,...} and
                        svm.pkg (one level up from svm/) which points into svm/
       (by having svm.pkg one level up we allow for a single site stata.toc with all packages,
        while keeping their contents separate; though Stata doesn't keep contents separate once installed)

For this commit, most of the worth is in svm/Makefile's "make svm.pkg",
which scans the directory and generates the file list in the form Stata can consume it,
with special casing for bin/ to handle cross-platformness. svm/Makefile will get rolled
into the main Makefile eventually. But first I need to re-read
[Recursive Make Considered Harmful](http://aegis.sourceforge.net/auug97.pdf)