Repair tests on Windows
I am not sure when they broke, but they sure were b0rkened.
Also, Stata installs updates and keeps the old exe around as
Stata-old.exe, which messes everything up because I makes $(wildcard)
return a list, but a list in make is the same as a list in bash:
whitespace separated tokens...which Windows already uses in its paths.