Windows fix
My tweaks to testing on Windows yesterday fell on deaf ears:
I had hard-coded the STATAPATH into my environment and didn't notice.
Apparently Windows treats trailing spaces in PATH as part of the content
of the last path, and we all know GNU make is fond of putting spaces
err'where.