_setenv.plugin
This was tricky. setenv() is not reliably cross-platform (and also an easy cause of security bugs).
I had to pull -std=c99 out of posix.mk because on Linux I need -std=gnu99 just to get the function.
It seems to behave, though, including that changes made with it are passed to subprocesses!

(the worst part of this commit was the 40 minutes I spent searching
 for what turned out to be a stray tab in the makefile)