Tidy the ereturn list
I use it to make svm_export safer. _svm.c::export() does check for no model,
however, before this you could do this:

```
svm_train a b c
regress a b c
svm_export using "awwyeah.model"
```
The export would export the SVM model from the first line,
not the OLS one from the second, which is at least very confusing.

The trouble is, of course, Stata's global-everything approach.
Without investing in learning their tacky OOP system, there's no way to encapsulate "this request goes belongs with this ...".

Maybe I could do do my own limited encapsulation: return the pointer (or some other equivalent ID number)
of the current svm_model as a Stata scalar, and then all the ops have to take it as a parameter (the `this` pointer!)
and if it is missing or changed they will know to bail.
..but it cannot change without a second call to _svm.c::train(), which already erases the old model,
so all it can do is go missing, which is effectively what checking for e(model)==svm does.
so nevermind.