First draft of manual page.
A lot is missing. I need to learn SVM (my course on it was not nearly enough).
Plus there's this nu-SVM variant which I think the libsvm people invented as
a research project.
And the libsvm README is no help: it defines things and says how to
call them but doesn't explain how to use them.

I also rearranged the arguments to train because I discovered that the
documentation flow was more meaningful in this different order.