Kludge: destroy probA and probB, just in case.
If i'm not going to rename probA->e(probA) (which is what return matrix = does, surprise! it's neither copy nor alias like it would be in every other language and in most other contexts in Stata including return scalar = ), then I should smash it to bits until I decide if I'm keeping it or not.