Document how to replicate results with svm-train.
On the (private) dataset Schonlau gave me to play with, prediction all ends up in the same category,
which is the category that 56% of the respondents are labelled. This seemed suspect, like a bug,
but with these commands I've replicated the identical result (and .model file), so if it's a bug
it's not in StataSVM. It's more likely a mixture of parameters and the overwhelming dominance of the one.
Maybe a different kernel would do better? RBF is for circles!

(and move Testing into CONTRIBUTING.md because COMPILE.md is possibly just for users
who are trying to get this built on an unusual platform)