Temporarily commit libsvm.dll (64 bit) as an ancillary file
This means that, on 64 bit Windows/Stata, you can do
. net install svm, from(http://..../)
. net get svm //grab libsvm.dll
. svm
and so long as you stay in the current directory (or, actually, so
long as you are in the download directory when you first run svm)
this will work.

When I get my magical bundling-DLLs trick working and stable, this should be removed.
(not the change to the makefile tho, that was an oversight which I only noticed
when I tried to use my ancillary files addition to makepluginpkg)