Replication of a libsvm bug
In short: svm_predict() and svm_predict_probability() give different results on certain datasets.

This is tricky. It only showed up on a dataset which we cannot share.
The attached data is an anonymized slice of the full dataset, and
luckily the bug still shows up there.
Committing to have a public record in case it heisenbugs on us.