Different aberrant library name algorithm
- first, try to load the aberrant name. if that succeeds, you win
- otherwise, fallback to ensurelib

we should *probably* start with the non-aberrant name and fallback to the aberrant one,
but then we get the error message "unable to load lib<name>.dll" which is probably confusing to end users.
If I could capture the error message output with a proper exception handler I could do it
or I could just replicate the error message manually.