Gridsearch example
This isn't a function because operating over arbitrary ranges of
parameters is really tricky. It would be tricky even in python,
but in Stata I have to worry about name collisions (eg what if a param matches a column that already exists)

Also allow folds(1) in cv.ado; it is a degenerate case, but it is nice to
be able to run it when gridsearching without having to change the rest of the code.