Repair click-to-run by using ado_from
click-to-run was broken if you did it in a second session after installing the package,
which is actually pretty common

. example svm binary_classification
Downloading svm ancillary files
file http://www.stata.com/svm.pkg not found
could not load svm.pkg from http://www.stata.com/
Unable to find binary_classification example.
r(3);

The trouble is that net looks in the previous "from" source,
which is stata.com if not set, rather than looking in where the
package was installed from.
So we just help it along a bit.

This works:
stata>
. net install svm, from(http://csclub.uwaterloo.ca/~nguenthe/stata)
. help svm
. // click-to-run

This doesn't:
stata>
. net install svm, from(http://csclub.uwaterloo.ca/~nguenthe/stata)
. exit
stata>
. help svm
. // click-to-run