Kill svm_{import,export}.
They don't get along with how Stata wants to work.

Even if I made svm_import reproduce most of the stored results that svm_train does,
there is no way to reproduce the set of variables used (e(depvar)/e(cmdline)) from
what's in libsvm .model file,.  At best, we could document this problem and provide
a command which lets the user *explicitly set* the variables, but that's a hack.

99% of our users will never ever do anything that isn't completely contained in a
single Stata .do file.  Saving models in Stata to reload them later is a perk but
definitely not necessary to have, and if we're honest they are probably more
confusing and stressful than useful for most users.

This code can always be retrieved from the git history if someone wants to take a
second crack at supporting this.