Merge ensurelib_aberrance into ensurelib.
The reason I had these as separate files was so that, *for the specific libraries that you
know have naming issues on Windows* like libsvm, you could say "ensurelib_aberrance <name>".
But I realized that it almost completely doesn't matter if we check for lib<name> first
because the chance of a library being named something like liblibname.so is so absurdly low.
To reduce the chance of a problem, this trick is still wrapped so that it only triggers on Windows,
but mostly it just won't.

And since I now know how to use inner functions in Stata, the two files can come together
in one beautiful whole.