Like the last commits, protect ensurelib by renaming it svm_ensurelib.
Actually, I took the old ensurelib and used it as an inner function inside of the new one, which is reasonably neat and tidy.
Now, ensurelib depends on _{dlopenable,{g,s}etenv}.plugin.
 I am not sure if I should rename these as well...

And I would have protected ensurepkg too, but it looks
like it's not actually used anywhere so it can go away.