New feature branch: export decision values
Decision values are what svm uses to make predictions.
In the case of regression, the "decision values" are defined
(by libsvm) to just be the regression values. But in one_class
and the classification models, the decision values are meaningful.
You can use them to tell how far from the hyperplane you are.

This is easy in principle, and libsvm already supports this:
 svm_predict_values instead of svm_predict()

but it's really quite tricky because
 a) we need to preallocate *a lower-triangular number* of spare columns, and make sure the numbers work out
 b) we need to be careful which columns are fed into the X vector in predict() and which are used as output
   (previously, there was only one column of output so I just hardcoded it first)
 c) we need to label the lower-triangular number of spare columns