"Extend" svm_model to track the dimension of the data.
This is the first refactoring to support decision values,
because I don't trust myself to always get the sizes right if I have to pass spare vars;
it's just more reliable to explicitly know the number of variables we're expecting.
(of course, we have no way to know if we get them in the *right order*...)

Also, parameterize stata2libsvm() to take an explicit region to work on,
which I noticed I could do when I was thinking about this.
and plan how to factor it better (in TODO.md)