Repair "predict missing to missing".
3a4348afa introduced this regression when it combined and cleaned up old code.
And, what with it being ancient code that I hadn't thought about in a while,
I assumed that predicting on missing data should just be disallowed,
and then I changed all the examples to match that.
But then I noticed the comment in binary_classification_example.do that explicitly
claims this is supported, so with some tweaking I've recovered the feature.