Major ONE_CLASS bugfix
Initially, I fixed svm_predict so that under one_class, it correctly operated on *all* variables instead of popping off the first one
 and then suddenly everything was being predicted to 0.
 Why would having *more* data make the results worse??

Well it turns out to be a case of two wrongs making a wrong that appeared to be a right.
Because of the order of operations---upper casing was happening too late---train
was incorrectly *not* duping the first variable:
 svm A B C, type(one_class)
 plugin call _svm A B C, "train" "ONE_CLASS"
which was *getting interpreted as*
 ignore A, train on (B,C)

Then svm_predict would re-extract "A B C", and incorrectly drop the first (under everything but ONE_CLASS this is correct), so
 predict P
 plugin call _svm P B C, "predict"
which would come out fine because it's using the same X vars in both cases
but the numbers, despite looking reasonable, are wrong because they're missing a whole dimension of data!

The reason this turned up when it did is that my change caused
 predict P
 plugin call _svm P A B C, "predict"
which meant it interpretedg As as Bs and Bs as Cs and I don't know what it's doing with the C...
ideally would have crashed but instead gave wrong answers. Hm.
side question:
 why wasn't this caught by my explicit dimension check?
 answer: the explicit dimension check that was added in 61da51f3 *only applies when probabilities are asked for*
 I've added a check in the case probabilities aren't asked for, too.

I never actually vetting this code path until just now, because I didn't have an interesting example to vet it on.