predict, scores would trigger an error not intended for it
The warning was intended for confirming that when you "predict, " svm_predict.ado passes
exactly the right number of variables to operate on. After that, a similar check would be
done for "predict, scores", outside the first if-else tree. Since I overlooked "return 1"
in that path, this turned into a warning that looked like an error, confusing the user.