Port example2smcl to Windows
I realized that if I could run make on Windows
I could also run awk just as easily. Should be
able to, anyhow.

svm_examples.ihlp is still in-repo for now,
but if this turns out to be feasible for all
of us to work with, I will add svm_examples.ihlp
to .gitignore.